IndexedDB: compress values before sending over IPC
Renderer compresses IDB values using snappy to save on IPC overhead and
to get the benefits of on-disk compression regardless of backing store.

There's no current intention to migrate existing values on disk.
This change is backwards compatible in that data coming from disk will
be interpreted appropriately regardless of whether it's compressed.

This change will be more useful when/if we migrate the backend from
leveldb to SQLite, but in the meantime it still serves a useful
purpose as noted in the linked bug.

Bug: 1459300
Change-Id: Idf036b7936314091bcf08cd30c92a671033893b2