Remove min overscroll event delta check
There doesn't seem to be spec describing how the deltas should be set
and Safari and Firefox don't seem to support the overscroll event yet
(per my manual test using this codepen[1] and wpt.fyi[2]) so there
doesn't really seem to be a reason to enforce a check on the minimum
delta in each and every overscroll event.
For example, before this patch, if there was extra scrolling input
of 300px causing overscroll events with deltas 48, 102 and 150 (which
sum up to 300) the test would fail.

It does seem reasonable to keep the check that the maximum overscroll
is reasonably close to and not greater than the excess amount of
scroll input.

[1]https://codepen.io/awogbemila/pen/eYQVWrL
[2]https://wpt.fyi/results/dom/events/scrolling/overscroll-deltas.html?label=experimental&label=master&aligned

Bug: 1430215
Change-Id: I891711f346d6f0cd8c3b423c0a2e8d141c6dec6a