Don't throw exceptions on unknown codecs or unsupported sizes.
Per spec isConfigSupported() should only throw a TypeError for
invalid configs. An unrecognized codec string doesn't mean the
config is invalid. Even a malformed codec string may just be
a newer standard of a given string.

Instead the spec says we should mark the codec as unsupported
when resolving the isConfigSupported() promise and throw a
NotSupportedError DOMException if that config is given to
configure().

This includes a couple of related changes:
- Chromium now handles missing displayWidth and displayHeight.
- Exceeding Chromium specific bounds reports unsupported.

Unfortunately the web tests weren't quite testing the behavior
required by the spec, so this updates them to do so.

Bug: 1457252
Change-Id: Id6b82dd09fdcb0d389429de15a6b23f6b6e5f56a