Fenced frames: Ensure that window.screen changes are not visible.
When window.screen changes in the top-level frame, the impact of that
can be seen in the nested fenced frame(s). We should ensure that only
the initial state of the top-level window.screen is visible to fenced
frames, so that alterations of values are not usable as a communication
channel.

Change-Id: If015f0be830b87a13716516edab0b4b11ea4d954