[no-vary-search-hint] Invalid NVS hint equals default URL variance
If No-Vary-Search hint is not specified as a string the speculation
rule will be ignored.

Otherwise, where there is an error in speculation rule parsing of
No-Vary-Search hint, don't discard the speculation rule, just use
default URL variance for No-Vary-Search hint instead. The behaviour
will be the same as if No-Vary-Search hint was not specified in the
first place.

Bug: 1457998
Change-Id: I76a86e53da8725ce56173014224b5719852fbd0d